+++ src/Model/Entity/Video.php
@@ -52,7 +52,7 @@ class Video extends Entity {
     protected $_accessible = [
         '*' => TRUE,
         'id' => FALSE,
-		'modified' => FALSE
+		'modified' => FALSE,
     ];
 	
 	/**